+++ metrics/web/pubspec.yaml
@@ -74,6 +74,7 @@ dev_dependencies:
   network_image_mock: ^1.0.2
   process_run: ^0.11.0+2
   test: any
+  uuid: ^2.2.2
 
 # For information on the generic Dart part of this file, see the
 # following page: https://dart.dev/tools/pub/pubspec
